Separate Egor from EgorSolver, use consuming builder
