+++ ego/src/egor_solver.rs
@@ -1,11 +1,19 @@
-//! Egor optimizer implements EGO algorithm with basic handling of constraints.
+//! Egor implementation as a [argmin::core::Solver] to be used to benefit from
+//! features coming with the argmin framework such as checkpointing or observers.
+//!
+//! Note: Depending on your need you can either use the `EgorSolver` or the provided
+//! `EgorBuilder` which allows to build an `Egor` struct which wraps the `argmin::Executor`
+//! running an `EgorSolver` on `ObjFun`. See [`EgorBuilder`]
 //!
 //! ```no_run
-//! # use ndarray::{array, Array2, ArrayView1, ArrayView2, Zip};
-//! # use egobox_doe::{Lhs, SamplingMethod};
-//! # use egobox_ego::{EgorBuilder, InfillStrategy, InfillOptimizer};
-//! # use rand_xoshiro::Xoshiro256Plus;
-//! # use ndarray_rand::rand::SeedableRng;
+//! use ndarray::{array, Array2, ArrayView1, ArrayView2, Zip};
+//! use egobox_doe::{Lhs, SamplingMethod};
+//! use egobox_ego::{EgorBuilder, InfillStrategy, InfillOptimizer, ObjFunc, EgorSolver};
+//! use egobox_moe::MoeParams;
+//! use rand_xoshiro::Xoshiro256Plus;
+//! use ndarray_rand::rand::SeedableRng;
+//! use argmin::core::Executor;
+//!
 //! use argmin_testfunctions::rosenbrock;
 //!
 //! // Rosenbrock test function: minimum y_opt = 0 at x_opt = (1, 1)
@@ -16,17 +24,15 @@
 //!         .par_for_each(|mut yi, xi| yi.assign(&array![rosenbrock(&xi.to_vec(), 1., 100.)]));
 //!     y
 //! }
-//!
+//! let rng = Xoshiro256Plus::seed_from_u64(42);
 //! let xlimits = array![[-2., 2.], [-2., 2.]];
-//! let res = EgorBuilder::optimize(rosenb)
-//!     .min_within(&xlimits)
-//!     .infill_strategy(InfillStrategy::EI)
-//!     .n_doe(10)
-//!     .target(1e-1)
-//!     .n_iter(30)
-//!     .run()
-//!     .expect("Rosenbrock minimization");
-//! println!("Rosenbrock min result = {:?}", res);
+//! let fobj = ObjFunc::new(rosenb);
+//! let solver: EgorSolver<MoeParams<f64, Xoshiro256Plus>> = EgorSolver::new(&xlimits, rng);
+//! let res = Executor::new(fobj, solver)
+//!             .configure(|state| state.max_iters(20))
+//!             .run()
+//!             .unwrap();
+//! println!("Rosenbrock min result = {:?}", res.state);
 //! ```
 //!
 //! Constraints are expected to be evaluated with the objective function
@@ -38,11 +44,13 @@
 //! if specified cstr values should be < `cstr_tol` (instead of < 0)
 //!
 //! ```no_run
-//! # use ndarray::{array, Array2, ArrayView1, ArrayView2, Zip};
-//! # use egobox_doe::{Lhs, SamplingMethod};
-//! # use egobox_ego::{EgorBuilder, InfillStrategy, InfillOptimizer};
-//! # use rand_xoshiro::Xoshiro256Plus;
-//! # use ndarray_rand::rand::SeedableRng;
+//! use ndarray::{array, Array2, ArrayView1, ArrayView2, Zip};
+//! use egobox_doe::{Lhs, SamplingMethod};
+//! use egobox_ego::{EgorBuilder, InfillStrategy, InfillOptimizer, ObjFunc, EgorSolver};
+//! use egobox_moe::MoeParams;
+//! use rand_xoshiro::Xoshiro256Plus;
+//! use ndarray_rand::rand::SeedableRng;
+//! use argmin::core::Executor;
 //!
 //! // Function G24: 1 global optimum y_opt = -5.5080 at x_opt =(2.3295, 3.1785)
 //! fn g24(x: &ArrayView1<f64>) -> f64 {
@@ -71,19 +79,24 @@
 //!     y
 //! }
 //!
+//! let rng = Xoshiro256Plus::seed_from_u64(42);
 //! let xlimits = array![[0., 3.], [0., 4.]];
 //! let doe = Lhs::new(&xlimits).sample(10);
-//! let res = EgorBuilder::optimize(f_g24)
-//!            .min_within(&xlimits)
-//!            .n_cstr(2)
-//!            .infill_strategy(InfillStrategy::EI)
-//!            .infill_optimizer(InfillOptimizer::Cobyla)
-//!            .doe(Some(doe))
-//!            .n_iter(40)
-//!            .target(-5.5080)
-//!            .run()
-//!            .expect("g24 minimized");
-//! println!("G24 min result = {:?}", res);
+//!
+//! let fobj = ObjFunc::new(f_g24);
+//! let solver: EgorSolver<MoeParams<f64, Xoshiro256Plus>> =
+//!   EgorSolver::new(&xlimits, rng)             
+//!     .n_cstr(2)
+//!     .infill_strategy(InfillStrategy::EI)
+//!     .infill_optimizer(InfillOptimizer::Cobyla)
+//!     .doe(Some(doe))
+//!     .target(-5.5080);
+//!
+//! let res = Executor::new(fobj, solver)
+//!             .configure(|state| state.max_iters(40))
+//!             .run()
+//!             .expect("g24 minimized");
+//! println!("G24 min result = {:?}", res.state);
 //! ```
 //!
 use crate::egor_state::{find_best_result_index, EgorState, MAX_POINT_ADDITION_RETRY};
@@ -112,14 +125,14 @@ use rand_xoshiro::Xoshiro256Plus;
 
 use argmin::argmin_error_closure;
 use argmin::core::{
-    CostFunction, Executor, Problem, Solver, State, TerminationReason, TerminationStatus, KV,
+    CostFunction, Problem, Solver, State, TerminationReason, TerminationStatus, KV,
 };
 
 use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 
-const DOE_INITIAL_FILE: &str = "egor_initial_doe.npy";
-const DOE_FILE: &str = "egor_doe.npy";
+pub const DOE_INITIAL_FILE: &str = "egor_initial_doe.npy";
+pub const DOE_FILE: &str = "egor_doe.npy";
 
 /// Implementation of `argmin::core::Solver` for Egor optimizer.
 /// Therefore this structure can be used with `argmin::core::Executor` and benefit
@@ -130,59 +143,59 @@ pub struct EgorSolver<SB: SurrogateBuilder> {
     /// Note 1 : The number of cost function evaluations is deduced using the following formula (n_doe + n_iter)
     /// Note 2 : When q_points > 1, the number of cost function evaluations is (n_doe + n_iter * q_points)
     /// is is an upper bounds as some points may be rejected as being to close to previous ones.   
-    n_iter: usize,
+    pub(crate) n_iter: usize,
     /// Number of starts for multistart approach used for hyperparameters optimization
-    n_start: usize,
+    pub(crate) n_start: usize,
     /// Number of points returned by EGO iteration (aka qEI Multipoint strategy)
     /// Actually as some point determination may fail (at most q_points are returned)
-    q_points: usize,
+    pub(crate) q_points: usize,
     /// Number of initial doe drawn using Latin hypercube sampling
     /// Note: n_doe > 0; otherwise n_doe = max(xdim + 1, 5)
-    n_doe: usize,
+    pub(crate) n_doe: usize,
     /// Number of Constraints
     /// Note: dim function ouput = 1 objective + n_cstr constraints
-    n_cstr: usize,
+    pub(crate) n_cstr: usize,
     /// Constraints violation tolerance meaning cstr < cstr_tol is considered valid
-    cstr_tol: f64,
+    pub(crate) cstr_tol: f64,
     /// Initial doe can be either \[x\] with x inputs only or an evaluated doe \[x, y\]
     /// Note: x dimension is determined using xlimits
-    doe: Option<Array2<f64>>,
+    pub(crate) doe: Option<Array2<f64>>,
     /// Multipoint strategy used to get several points to be evaluated at each iteration
-    q_ei: QEiStrategy,
+    pub(crate) q_ei: QEiStrategy,
     /// Criterium to select next point to evaluate
-    infill: InfillStrategy,
+    pub(crate) infill: InfillStrategy,
     /// The optimizer used to optimize infill criterium
-    infill_optimizer: InfillOptimizer,
+    pub(crate) infill_optimizer: InfillOptimizer,
     /// Regression specification for GP models used by mixture of experts (see [egobox_moe])
-    regression_spec: RegressionSpec,
+    pub(crate) regression_spec: RegressionSpec,
     /// Correlation specification for GP models used by mixture of experts (see [egobox_moe])
-    correlation_spec: CorrelationSpec,
+    pub(crate) correlation_spec: CorrelationSpec,
     /// Optional dimension reduction (see [egobox_moe])
-    kpls_dim: Option<usize>,
+    pub(crate) kpls_dim: Option<usize>,
     /// Number of clusters used by mixture of experts (see [egobox_moe])
     /// When set to 0 the clusters are computes automatically and refreshed
     /// every 10-points (tentative) additions
-    n_clusters: Option<usize>,
+    pub(crate) n_clusters: Option<usize>,
     /// Specification of a target objective value which is used to stop the algorithm once reached
-    target: f64,
+    pub(crate) target: f64,
     /// Directory to save intermediate results: inital doe + evalutions at each iteration
-    outdir: Option<String>,
+    pub(crate) outdir: Option<String>,
     /// If true use `outdir` to retrieve and start from previous results
-    hot_start: bool,
+    pub(crate) hot_start: bool,
     /// Matrix (nx, 2) of [lower bound, upper bound] of the nx components of x
     /// Note: used for continuous variables handling, the optimizer base.
-    xlimits: Array2<f64>,
+    pub(crate) xlimits: Array2<f64>,
     /// List of x types allowing the handling of discrete input variables
-    xtypes: Option<Vec<Xtype>>,
+    pub(crate) xtypes: Option<Vec<Xtype>>,
     /// Flag for discrete handling, true if mixed-integer type present in xtypes, otherwise false
-    no_discrete: bool,
+    pub(crate) no_discrete: bool,
     /// An optional surrogate builder used to model objective and constraint
     /// functions, otherwise [mixture of expert](egobox_moe) is used
     /// Note: if specified takes precedence over individual settings
-    surrogate_builder: SB,
+    pub(crate) surrogate_builder: SB,
     /// A random generator used to get reproductible results.
     /// For instance: Xoshiro256Plus::from_u64_seed(42) for reproducibility
-    rng: Xoshiro256Plus,
+    pub(crate) rng: Xoshiro256Plus,
 }
 
 impl SurrogateBuilder for MoeParams<f64, Xoshiro256Plus> {
@@ -312,19 +325,19 @@ impl<SB: SurrogateBuilder> EgorSolver<SB> {
     }
 
     /// Sets allowed number of evaluation of the function under optimization
-    pub fn n_iter(&mut self, n_iter: usize) -> &mut Self {
+    pub fn n_iter(mut self, n_iter: usize) -> Self {
         self.n_iter = n_iter;
         self
     }
 
     /// Sets the number of runs of infill strategy optimizations (best result taken)
-    pub fn n_start(&mut self, n_start: usize) -> &mut Self {
+    pub fn n_start(mut self, n_start: usize) -> Self {
         self.n_start = n_start;
         self
     }
 
     /// Sets Number of parallel evaluations of the function under optimization
-    pub fn q_points(&mut self, q_points: usize) -> &mut Self {
+    pub fn q_points(mut self, q_points: usize) -> Self {
         self.q_points = q_points;
         self
     }
@@ -333,19 +346,19 @@ impl<SB: SurrogateBuilder> EgorSolver<SB> {
     ///
     /// When 0 a number of points is computed automatically regarding the number of input variables
     /// of the function under optimization.
-    pub fn n_doe(&mut self, n_doe: usize) -> &mut Self {
+    pub fn n_doe(mut self, n_doe: usize) -> Self {
         self.n_doe = n_doe;
         self
     }
 
     /// Sets the number of constraint functions
-    pub fn n_cstr(&mut self, n_cstr: usize) -> &mut Self {
+    pub fn n_cstr(mut self, n_cstr: usize) -> Self {
         self.n_cstr = n_cstr;
         self
     }
 
     /// Sets the tolerance on constraints violation (cstr < tol)
-    pub fn cstr_tol(&mut self, tol: f64) -> &mut Self {
+    pub fn cstr_tol(mut self, tol: f64) -> Self {
         self.cstr_tol = tol;
         self
     }
@@ -354,7 +367,7 @@ impl<SB: SurrogateBuilder> EgorSolver<SB> {
     ///
     /// Either nt = nx then only x are specified and ns evals are done to get y doe values,
     /// or nt = nx + ny then x = doe(:, :nx) and y = doe(:, nx:) are specified
-    pub fn doe(&mut self, doe: Option<Array2<f64>>) -> &mut Self {
+    pub fn doe(mut self, doe: Option<Array2<f64>>) -> Self {
         self.doe = doe.map(|x| x.to_owned());
         self
     }
@@ -363,31 +376,31 @@ impl<SB: SurrogateBuilder> EgorSolver<SB> {
     ///
     /// Parallel infill criterion to get virtual next promising points in order to allow
     /// n parallel evaluations of the function under optimization.
-    pub fn qei_strategy(&mut self, q_ei: QEiStrategy) -> &mut Self {
+    pub fn qei_strategy(mut self, q_ei: QEiStrategy) -> Self {
         self.q_ei = q_ei;
         self
     }
 
     /// Sets the nfill criteria
-    pub fn infill_strategy(&mut self, infill: InfillStrategy) -> &mut Self {
+    pub fn infill_strategy(mut self, infill: InfillStrategy) -> Self {
         self.infill = infill;
         self
     }
 
     /// Sets the infill optimizer
-    pub fn infill_optimizer(&mut self, optimizer: InfillOptimizer) -> &mut Self {
+    pub fn infill_optimizer(mut self, optimizer: InfillOptimizer) -> Self {
         self.infill_optimizer = optimizer;
         self
     }
 
     /// Sets the allowed regression models used in gaussian processes.
-    pub fn regression_spec(&mut self, regression_spec: RegressionSpec) -> &mut Self {
+    pub fn regression_spec(mut self, regression_spec: RegressionSpec) -> Self {
         self.regression_spec = regression_spec;
         self
     }
 
     /// Sets the allowed correlation models used in gaussian processes.
-    pub fn correlation_spec(&mut self, correlation_spec: CorrelationSpec) -> &mut Self {
+    pub fn correlation_spec(mut self, correlation_spec: CorrelationSpec) -> Self {
         self.correlation_spec = correlation_spec;
         self
     }
@@ -395,7 +408,7 @@ impl<SB: SurrogateBuilder> EgorSolver<SB> {
     /// Sets the number of components to be used specifiying PLS projection is used (a.k.a KPLS method).
     ///
     /// This is used to address high-dimensional problems typically when nx > 9.
-    pub fn kpls_dim(&mut self, kpls_dim: Option<usize>) -> &mut Self {
+    pub fn kpls_dim(mut self, kpls_dim: Option<usize>) -> Self {
         self.kpls_dim = kpls_dim;
         self
     }
@@ -404,25 +417,25 @@ impl<SB: SurrogateBuilder> EgorSolver<SB> {
     ///
     /// When set to Some(0), the number of clusters is determined automatically
     /// When set None, default to 1
-    pub fn n_clusters(&mut self, n_clusters: Option<usize>) -> &mut Self {
+    pub fn n_clusters(mut self, n_clusters: Option<usize>) -> Self {
         self.n_clusters = n_clusters;
         self
     }
 
     /// Sets a known target minimum to be used as a stopping criterion.
-    pub fn target(&mut self, target: f64) -> &mut Self {
+    pub fn target(mut self, target: f64) -> Self {
         self.target = target;
         self
     }
 
     /// Sets a directory to write optimization history and used as search path for hot start doe
-    pub fn outdir(&mut self, outdir: Option<String>) -> &mut Self {
+    pub fn outdir(mut self, outdir: Option<String>) -> Self {
         self.outdir = outdir;
         self
     }
 
     /// Whether we start by loading last DOE saved in `outdir` as initial DOE
-    pub fn hot_start(&mut self, hot_start: bool) -> &mut Self {
+    pub fn hot_start(mut self, hot_start: bool) -> Self {
         self.hot_start = hot_start;
         self
     }
@@ -1197,363 +1210,20 @@ where
     }
 }
 
-/// EGO optimizer builder allowing to specify function to be minimized
-/// subject to constraints intended to be negative.
-///
-pub struct EgorBuilder<O: GroupFunc> {
-    fobj: O,
-    seed: Option<u64>,
-}
-
-impl<O: GroupFunc> EgorBuilder<O> {
-    /// Function to be minimized domain should be basically R^nx -> R^ny
-    /// where nx is the dimension of input x and ny the output dimension
-    /// equal to 1 (obj) + n (cstrs).
-    /// But function has to be able to evaluate several points in one go
-    /// hence take an (p, nx) matrix and return an (p, ny) matrix
-    pub fn optimize(fobj: O) -> Self {
-        EgorBuilder { fobj, seed: None }
-    }
-
-    /// Allow to specify a seed for random number generator to allow
-    /// reproducible runs.
-    pub fn random_seed(mut self, seed: u64) -> Self {
-        self.seed = Some(seed);
-        self
-    }
-
-    /// Build an Egor optimizer to minimize the function within
-    /// the continuous xlimits specified as [[lower, upper], ...]
-    /// number of rows gives the dimension of the inputs (continuous optimization).
-    pub fn min_within(
-        self,
-        xlimits: &ArrayBase<impl Data<Elem = f64>, Ix2>,
-    ) -> Egor<O, MoeParams<f64, Xoshiro256Plus>> {
-        let rng = if let Some(seed) = self.seed {
-            Xoshiro256Plus::seed_from_u64(seed)
-        } else {
-            Xoshiro256Plus::from_entropy()
-        };
-        Egor {
-            fobj: ObjFunc::new(self.fobj),
-            solver: EgorSolver::new(xlimits, rng),
-        }
-    }
-
-    /// Build an Egor optimizer to minimize the function R^n -> R^p taking
-    /// inputs specified with given xtypes where some of components may be
-    /// discrete variables (mixed-integer optimization).
-    pub fn min_within_mixed_space(self, xtypes: &[Xtype]) -> Egor<O, MixintMoeParams> {
-        let rng = if let Some(seed) = self.seed {
-            Xoshiro256Plus::seed_from_u64(seed)
-        } else {
-            Xoshiro256Plus::from_entropy()
-        };
-        Egor {
-            fobj: ObjFunc::new(self.fobj),
-            solver: EgorSolver::new_with_xtypes(xtypes, rng),
-        }
-    }
-}
-
-/// Egor optimizer structure used to parameterize the underlying `argmin::Solver`
-/// and trigger the optimization using `argmin::Executor`.
-#[derive(Clone)]
-pub struct Egor<O: GroupFunc, SB: SurrogateBuilder> {
-    fobj: ObjFunc<O>,
-    solver: EgorSolver<SB>,
-}
-
-impl<O: GroupFunc, SB: SurrogateBuilder> Egor<O, SB> {
-    /// Sets allowed number of evaluation of the function under optimization
-    pub fn n_iter(&mut self, n_iter: usize) -> &mut Self {
-        self.solver.n_iter(n_iter);
-        self
-    }
-
-    /// Sets the number of runs of infill strategy optimizations (best result taken)
-    pub fn n_start(&mut self, n_start: usize) -> &mut Self {
-        self.solver.n_start(n_start);
-        self
-    }
-
-    /// Sets Number of parallel evaluations of the function under optimization
-    pub fn q_points(&mut self, q_points: usize) -> &mut Self {
-        self.solver.q_points(q_points);
-        self
-    }
-
-    /// Number of samples of initial LHS sampling (used when DOE not provided by the user)
-    ///
-    /// When 0 a number of points is computed automatically regarding the number of input variables
-    /// of the function under optimization.
-    pub fn n_doe(&mut self, n_doe: usize) -> &mut Self {
-        self.solver.n_doe(n_doe);
-        self
-    }
-
-    /// Sets the number of constraint functions
-    pub fn n_cstr(&mut self, n_cstr: usize) -> &mut Self {
-        self.solver.n_cstr(n_cstr);
-        self
-    }
-
-    /// Sets the tolerance on constraints violation (cstr < tol)
-    pub fn cstr_tol(&mut self, tol: f64) -> &mut Self {
-        self.solver.cstr_tol(tol);
-        self
-    }
-
-    /// Sets an initial DOE containing ns samples
-    ///
-    /// Either nt = nx then only x are specified and ns evals are done to get y doe values,
-    /// or nt = nx + ny then x = doe(:, :nx) and y = doe(:, nx:) are specified
-    pub fn doe(&mut self, doe: Option<Array2<f64>>) -> &mut Self {
-        self.solver.doe(doe);
-        self
-    }
-
-    /// Sets the parallel infill strategy
-    ///
-    /// Parallel infill criteria to get virtual next promising points in order to allow
-    /// n parallel evaluations of the function under optimization.
-    pub fn qei_strategy(&mut self, q_ei: QEiStrategy) -> &mut Self {
-        self.solver.qei_strategy(q_ei);
-        self
-    }
-
-    /// Sets the nfill criteria
-    pub fn infill_strategy(&mut self, infill: InfillStrategy) -> &mut Self {
-        self.solver.infill_strategy(infill);
-        self
-    }
-
-    /// Sets the infill optimizer
-    pub fn infill_optimizer(&mut self, optimizer: InfillOptimizer) -> &mut Self {
-        self.solver.infill_optimizer(optimizer);
-        self
-    }
-
-    /// Sets the allowed regression models used in gaussian processes.
-    pub fn regression_spec(&mut self, regression_spec: RegressionSpec) -> &mut Self {
-        self.solver.regression_spec(regression_spec);
-        self
-    }
-
-    /// Sets the allowed correlation models used in gaussian processes.
-    pub fn correlation_spec(&mut self, correlation_spec: CorrelationSpec) -> &mut Self {
-        self.solver.correlation_spec(correlation_spec);
-        self
-    }
-
-    /// Sets the number of components to be used specifiying PLS projection is used (a.k.a KPLS method).
-    ///
-    /// This is used to address high-dimensional problems typically when nx > 9.
-    pub fn kpls_dim(&mut self, kpls_dim: Option<usize>) -> &mut Self {
-        self.solver.kpls_dim(kpls_dim);
-        self
-    }
-
-    /// Sets the number of clusters used by the mixture of surrogate experts.
-    ///
-    /// When set to 0, the number of clusters is determined automatically
-    pub fn n_clusters(&mut self, n_clusters: Option<usize>) -> &mut Self {
-        self.solver.n_clusters(n_clusters);
-        self
-    }
-
-    /// Sets a known target minimum to be used as a stopping criterion.
-    pub fn target(&mut self, target: f64) -> &mut Self {
-        self.solver.target(target);
-        self
-    }
-
-    /// Sets a directory to write optimization history and used as search path for hot start doe
-    pub fn outdir(&mut self, outdir: Option<String>) -> &mut Self {
-        self.solver.outdir(outdir);
-        self
-    }
-
-    /// Whether we start by loading last DOE saved in `outdir` as initial DOE
-    pub fn hot_start(&mut self, hot_start: bool) -> &mut Self {
-        self.solver.hot_start(hot_start);
-        self
-    }
-
-    /// Given an evaluated doe (x, y) data, return the next promising x point
-    /// where optimum may occurs regarding the infill criterium.
-    /// This function inverse the control of the optimization and can used
-    /// ask-and-tell interface to the EGO optmizer.
-    pub fn suggest(
-        &self,
-        x_data: &ArrayBase<impl Data<Elem = f64>, Ix2>,
-        y_data: &ArrayBase<impl Data<Elem = f64>, Ix2>,
-    ) -> Array2<f64> {
-        self.solver.suggest(x_data, y_data)
-    }
-
-    /// Runs the (constrained) optimization of the objective function.
-    pub fn run(&self) -> Result<OptimResult<f64>> {
-        let no_discrete = self.solver.no_discrete;
-        let xtypes = self.solver.xtypes.clone();
-
-        let result = Executor::new(self.fobj.clone(), self.solver.clone()).run()?;
-        debug!("ARGMIN result = {}", result);
-        let (x_data, y_data) = result.state().clone().take_data().unwrap();
-
-        let res = if no_discrete {
-            info!("History: \n{}", concatenate![Axis(1), x_data, y_data]);
-            OptimResult {
-                x_opt: result.state.get_best_param().unwrap().to_owned(),
-                y_opt: result.state.get_full_best_cost().unwrap().to_owned(),
-                x_hist: x_data,
-                y_hist: y_data,
-            }
-        } else {
-            let xtypes = xtypes.unwrap(); // !no_discrete
-            let x_data = cast_to_discrete_values(&xtypes, &x_data);
-            let x_data = fold_with_enum_index(&xtypes, &x_data.view());
-            info!("History: \n{}", concatenate![Axis(1), x_data, y_data]);
-
-            let x_opt = result
-                .state
-                .get_best_param()
-                .unwrap()
-                .to_owned()
-                .insert_axis(Axis(0));
-            let x_opt = cast_to_discrete_values(&xtypes, &x_opt);
-            let x_opt = fold_with_enum_index(&xtypes, &x_opt.view());
-            OptimResult {
-                x_opt: x_opt.row(0).to_owned(),
-                y_opt: result.state.get_full_best_cost().unwrap().to_owned(),
-                x_hist: x_data,
-                y_hist: y_data,
-            }
-        };
-        info!("Optim Result: min f(x)={} at x={}", res.y_opt, res.x_opt);
-
-        Ok(res)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use approx::assert_abs_diff_eq;
+    use argmin::core::Executor;
     use argmin_testfunctions::rosenbrock;
-    use ndarray::{array, ArrayView2};
-    use ndarray_npy::read_npy;
-    use serial_test::serial;
-    use std::time::Instant;
-
-    #[cfg(not(feature = "blas"))]
-    use linfa_linalg::norm::*;
-    #[cfg(feature = "blas")]
-    use ndarray_linalg::Norm;
-
-    fn xsinx(x: &ArrayView2<f64>) -> Array2<f64> {
-        (x - 3.5) * ((x - 3.5) / std::f64::consts::PI).mapv(|v| v.sin())
-    }
+    use ndarray::{array, ArrayView1};
 
     #[test]
-    #[serial]
-    fn test_xsinx_ei_quadratic_egor_solver() {
-        let initial_doe = array![[0.], [7.], [25.]];
-        let res = EgorBuilder::optimize(xsinx)
-            .min_within(&array![[0.0, 25.0]])
-            .infill_strategy(InfillStrategy::EI)
-            .regression_spec(RegressionSpec::QUADRATIC)
-            .correlation_spec(CorrelationSpec::ALL)
-            .n_iter(30)
-            .doe(Some(initial_doe.to_owned()))
-            .target(-15.1)
-            .outdir(Some("target/tests".to_string()))
-            .run()
-            .expect("Egor should minimize xsinx");
-        let expected = array![-15.1];
-        assert_abs_diff_eq!(expected, res.y_opt, epsilon = 0.5);
-        let saved_doe: Array2<f64> = read_npy(format!("target/tests/{DOE_INITIAL_FILE}")).unwrap();
-        assert_abs_diff_eq!(initial_doe, saved_doe.slice(s![..3, ..1]), epsilon = 1e-6);
-    }
-
-    #[test]
-    #[serial]
-    fn test_xsinx_wb2_egor_solver() {
-        let res = EgorBuilder::optimize(xsinx)
-            .min_within(&array![[0.0, 25.0]])
-            .n_iter(20)
-            .regression_spec(RegressionSpec::ALL)
-            .correlation_spec(CorrelationSpec::ALL)
-            .run()
-            .expect("Egor should minimize");
-        let expected = array![18.9];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 1e-1);
-    }
-
-    #[test]
-    #[serial]
-    fn test_xsinx_auto_clustering_egor_solver() {
-        let res = EgorBuilder::optimize(xsinx)
-            .min_within(&array![[0.0, 25.0]])
-            .n_clusters(Some(0))
-            .n_iter(20)
-            .run()
-            .expect("Egor with auto clustering should minimize xsinx");
-        let expected = array![18.9];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 1e-1);
-    }
-
-    #[test]
-    #[serial]
-    fn test_xsinx_with_hotstart_egor_solver() {
-        let xlimits = array![[0.0, 25.0]];
-        let doe = Lhs::new(&xlimits).sample(10);
-        let res = EgorBuilder::optimize(xsinx)
-            .random_seed(42)
-            .min_within(&xlimits)
-            .n_iter(15)
-            .doe(Some(doe))
-            .outdir(Some("target/tests".to_string()))
-            .run()
-            .expect("Minimize failure");
-        let expected = array![18.9];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 1e-1);
-
-        let res = EgorBuilder::optimize(xsinx)
-            .random_seed(42)
-            .min_within(&xlimits)
-            .n_iter(5)
-            .outdir(Some("target/tests".to_string()))
-            .hot_start(true)
-            .run()
-            .expect("Egor should minimize xsinx");
-        let expected = array![18.9];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 1e-1);
-    }
-
-    #[test]
-    #[serial]
-    fn test_xsinx_suggestions_egor_solver() {
-        let mut ego = EgorBuilder::optimize(xsinx)
-            .random_seed(42)
-            .min_within(&array![[0., 25.]]);
-        let ego = ego
-            .regression_spec(RegressionSpec::ALL)
-            .correlation_spec(CorrelationSpec::ALL)
-            .infill_strategy(InfillStrategy::EI);
-
-        let mut doe = array![[0.], [7.], [20.], [25.]];
-        let mut y_doe = xsinx(&doe.view());
-        for _i in 0..10 {
-            let x_suggested = ego.suggest(&doe, &y_doe);
-
-            doe = concatenate![Axis(0), doe, x_suggested];
-            y_doe = xsinx(&doe.view());
-        }
-
-        let expected = -15.1;
-        let y_opt = y_doe.min().unwrap();
-        assert_abs_diff_eq!(expected, *y_opt, epsilon = 1e-1);
+    fn test_unfold_xtypes_as_continuous_limits() {
+        let xtypes = vec![Xtype::Int(0, 25)];
+        let xlimits = unfold_xtypes_as_continuous_limits(&xtypes);
+        let expected = array![[0., 25.]];
+        assert_abs_diff_eq!(expected, xlimits);
     }
 
     fn rosenb(x: &ArrayView2<f64>) -> Array2<f64> {
@@ -1565,45 +1235,39 @@ mod tests {
     }
 
     #[test]
-    #[serial]
-    fn test_rosenbrock_2d_egor_solver() {
-        let now = Instant::now();
+    fn test_rosenbrock_egor_solver() {
+        let rng = Xoshiro256Plus::seed_from_u64(42);
         let xlimits = array![[-2., 2.], [-2., 2.]];
-        let doe = Lhs::new(&xlimits)
-            .with_rng(Xoshiro256Plus::seed_from_u64(42))
-            .sample(10);
-        let res = EgorBuilder::optimize(rosenb)
-            .random_seed(42)
-            .min_within(&xlimits)
-            .doe(Some(doe))
-            .n_iter(100)
-            .regression_spec(RegressionSpec::ALL)
-            .correlation_spec(CorrelationSpec::ALL)
-            .target(1e-2)
+
+        let fobj = ObjFunc::new(rosenb);
+        let solver: EgorSolver<MoeParams<f64, Xoshiro256Plus>> = EgorSolver::new(&xlimits, rng);
+
+        let res = Executor::new(fobj, solver)
+            .configure(|state| state.max_iters(20))
             .run()
-            .expect("Minimize failure");
-        println!("Rosenbrock optim result = {res:?}");
-        println!("Elapsed = {:?}", now.elapsed());
+            .unwrap();
+
         let expected = array![1., 1.];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 5e-1);
+        assert_abs_diff_eq!(
+            expected,
+            res.state.get_best_param().unwrap(),
+            epsilon = 5e-1
+        );
     }
 
-    // Objective
-    fn g24(x: &ArrayBase<impl Data<Elem = f64>, Ix1>) -> f64 {
-        // Function G24: 1 global optimum y_opt = -5.5080 at x_opt =(2.3295, 3.1785)
+    // Function G24: 1 global optimum y_opt = -5.5080 at x_opt =(2.3295, 3.1785)
+    fn g24(x: &ArrayView1<f64>) -> f64 {
         -x[0] - x[1]
     }
-
     // Constraints < 0
-    fn g24_c1(x: &ArrayBase<impl Data<Elem = f64>, Ix1>) -> f64 {
+    fn g24_c1(x: &ArrayView1<f64>) -> f64 {
         -2.0 * x[0].powf(4.0) + 8.0 * x[0].powf(3.0) - 8.0 * x[0].powf(2.0) + x[1] - 2.0
     }
-
-    fn g24_c2(x: &ArrayBase<impl Data<Elem = f64>, Ix1>) -> f64 {
+    fn g24_c2(x: &ArrayView1<f64>) -> f64 {
         -4.0 * x[0].powf(4.0) + 32.0 * x[0].powf(3.0) - 88.0 * x[0].powf(2.0) + 96.0 * x[0] + x[1]
             - 36.0
     }
-
+    // Gouped function : objective + constraints
     fn f_g24(x: &ArrayView2<f64>) -> Array2<f64> {
         let mut y = Array2::zeros((x.nrows(), 3));
         Zip::from(y.rows_mut())
@@ -1615,119 +1279,26 @@ mod tests {
     }
 
     #[test]
-    #[serial]
-    fn test_egor_g24_basic_egor_solver() {
+    fn test_g24_egor_solver() {
+        let rng = Xoshiro256Plus::seed_from_u64(42);
         let xlimits = array![[0., 3.], [0., 4.]];
-        let doe = Lhs::new(&xlimits)
-            .with_rng(Xoshiro256Plus::seed_from_u64(42))
-            .sample(3);
-        let res = EgorBuilder::optimize(f_g24)
-            .random_seed(42)
-            .min_within(&xlimits)
-            .n_cstr(2)
-            .doe(Some(doe))
-            .n_iter(20)
-            .run()
-            .expect("Minimize failure");
-        println!("G24 optim result = {res:?}");
-        let expected = array![2.3295, 3.1785];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 2e-2);
-    }
+        let fobj = ObjFunc::new(f_g24);
+        let solver = EgorSolver::<MoeParams<f64, _>>::new(&xlimits, rng)
+            .n_doe(3)
+            .n_cstr(2);
 
-    #[test]
-    fn test_egor_g24_qei_egor_solver() {
-        let xlimits = array![[0., 3.], [0., 4.]];
-        let doe = Lhs::new(&xlimits)
-            .with_rng(Xoshiro256Plus::seed_from_u64(42))
-            .sample(10);
-        let res = EgorBuilder::optimize(f_g24)
-            .random_seed(42)
-            .min_within(&xlimits)
-            .regression_spec(RegressionSpec::ALL)
-            .correlation_spec(CorrelationSpec::ALL)
-            .n_cstr(2)
-            .q_points(2)
-            .qei_strategy(QEiStrategy::KrigingBeliever)
-            .doe(Some(doe))
-            .target(-5.508013)
-            .n_iter(30)
-            .run()
-            .expect("Egor minimization");
-        println!("G24 optim result = {res:?}");
-        let expected = array![2.3295, 3.1785];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 2e-2);
-    }
-
-    // Mixed-integer tests
-
-    fn mixsinx(x: &ArrayView2<f64>) -> Array2<f64> {
-        if (x.mapv(|v| v.round()).norm_l2() - x.norm_l2()).abs() < 1e-6 {
-            (x - 3.5) * ((x - 3.5) / std::f64::consts::PI).mapv(|v| v.sin())
-        } else {
-            panic!("Error: mixsinx works only on integer, got {:?}", x)
-        }
-    }
-
-    #[test]
-    #[serial]
-    fn test_mixsinx_ei_mixint_egor_solver() {
-        let n_iter = 30;
-        let doe = array![[0.], [7.], [25.]];
-        let xtypes = vec![Xtype::Int(0, 25)];
-
-        let res = EgorBuilder::optimize(mixsinx)
-            .random_seed(42)
-            .min_within_mixed_space(&xtypes)
-            .doe(Some(doe))
-            .n_iter(n_iter)
-            .target(-15.1)
-            .infill_strategy(InfillStrategy::EI)
+        let res = Executor::new(fobj, solver)
+            .configure(|state| state.max_iters(20))
             .run()
             .unwrap();
-        assert_abs_diff_eq!(array![18.], res.x_opt, epsilon = 2.);
-    }
 
-    #[test]
-    fn test_mixsinx_reclustering_mixint_egor_solver() {
-        let n_iter = 30;
-        let doe = array![[0.], [7.], [25.]];
-        let xtypes = vec![Xtype::Int(0, 25)];
-
-        let res = EgorBuilder::optimize(mixsinx)
-            .random_seed(42)
-            .min_within_mixed_space(&xtypes)
-            .doe(Some(doe))
-            .n_iter(n_iter)
-            .target(-15.1)
-            .infill_strategy(InfillStrategy::EI)
-            .run()
-            .unwrap();
-        assert_abs_diff_eq!(array![18.], res.x_opt, epsilon = 2.);
-    }
-
-    #[test]
-    #[serial]
-    fn test_mixsinx_wb2_mixint_egor_solver() {
-        let n_iter = 30;
-        let xtypes = vec![Xtype::Int(0, 25)];
-
-        let res = EgorBuilder::optimize(mixsinx)
-            .random_seed(42)
-            .min_within_mixed_space(&xtypes)
-            .regression_spec(egobox_moe::RegressionSpec::CONSTANT)
-            .correlation_spec(egobox_moe::CorrelationSpec::SQUAREDEXPONENTIAL)
-            .n_iter(n_iter)
-            .infill_strategy(InfillStrategy::WB2)
-            .run()
-            .unwrap();
-        assert_abs_diff_eq!(&array![18.], &res.x_opt, epsilon = 3.);
-    }
-
-    #[test]
-    fn test_unfold_xtypes_as_continuous_limits() {
-        let xtypes = vec![Xtype::Int(0, 25)];
-        let xlimits = unfold_xtypes_as_continuous_limits(&xtypes);
-        let expected = array![[0., 25.]];
-        assert_abs_diff_eq!(expected, xlimits);
+        let expected = array![2.3295201833653514, 3.178493151673985];
+        assert_abs_diff_eq!(
+            expected,
+            res.state.get_best_param().unwrap(),
+            epsilon = 5e-1
+        );
+        let expected = -5.50;
+        assert_abs_diff_eq!(expected, res.state.get_best_cost(), epsilon = 5e-1);
     }
 }
